requirements.txt
# Core Framework
crewai>=0.28.0

# Data Validation
pydantic>=2.0.0

# Vector Database
chromadb>=0.4.0
pinecone-client>=2.2.0

# LLM Integration
anthropic>=0.7.0
openai>=1.0.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.20.0
streamlit>=1.25.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.28.0

# Graph Database (for GraphRAG)
neo4j>=5.0.0

# Vector Database
sentence-transformers>=2.2.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0

# Discord Bot
discord.py>=2.3.0
